+++ src/components/ui/table.tsx
@@ -88,7 +88,7 @@ const TableCell = React.forwardRef<
   <td
     ref={ref}
     className={cn(
-      "p-2 sm:p-3 md:p-4 lg:p-4 align-middle [&:has([role=checkbox])]:pr-0",
+      "p-[6px] sm:p-3 md:p-4 lg:p-4 align-middle [&:has([role=checkbox])]:pr-0",
       className
     )}
     {...props}
