+++ src/pages/home.tsx
@@ -1,21 +1,13 @@
 import { NextPage } from "next";
 
 import { FaHome } from "react-icons/fa";
-import {
-  Table,
-  TableBody,
-  TableCaption,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from "@/components/ui/table";
 import * as React from "react";
 import { useRouter } from "next/router";
 import withAuth from "@/components/authorise";
 import Menu from "@/components/menu";
+import HomeTable from "@/components/home/homeTable";
 
-export type TradeData = {
+export type HomeData = {
   accountid: number;
   totalTrades: number;
   totalWins: number;
@@ -26,7 +18,7 @@ export type TradeData = {
 
 const Home: NextPage = () => {
   const router = useRouter();
-  const [data, setData] = React.useState<TradeData[]>([]);
+  const [data, setData] = React.useState<HomeData[]>([]);
   const [menuOpen, setMenuOpen] = React.useState(false);
   const [user, setUser] = React.useState<string>("");
   React.useEffect(() => {
@@ -48,7 +40,7 @@ const Home: NextPage = () => {
               body: JSON.stringify(loggedInVal),
               headers: { "Content-Type": "application/json" },
             });
-            const tradeData: TradeData[] = await response.json();
+            const tradeData: HomeData[] = await response.json();
             console.log(tradeData);
             setData(tradeData);
           }
@@ -70,77 +62,13 @@ const Home: NextPage = () => {
             <FaHome className="h-10 w-10"></FaHome>
             <span className="my-auto font-bold">Home</span>
           </div>
-          <div className="p-2 mt-5 mx-5 text-center">
+          <div className="p-2 mt-5 mx-3 text-center">
             <h1 className="text-2xl sm:text-2xl md:text-3xl lg:text-4xl font-extrabold">
               Welcome {user ? user : "..."}
             </h1>
           </div>
-          <div className="flex flex-1 overflow-auto p-1 justify-center mx-auto">
-            {user ? (
-              <Table className="bg-gray-400 w-[300px] sm:w-[400px] md:w-[500px] lg:w-[700px] rounded-2xl">
-                <TableCaption className="text-gray-500">
-                  A Table of your recent Trades.
-                </TableCaption>
-                <TableHeader>
-                  <TableRow>
-                    <TableHead className="text-slate-200 text-xs lg:text-2xl md:text-lg sm:text-sm">
-                      Account ID
-                    </TableHead>
-                    <TableHead className="text-slate-200 text-xs lg:text-2xl md:text-lg sm:text-sm">
-                      N<sup>o</sup> Trades
-                    </TableHead>
-                    <TableHead className="text-slate-200 text-xs lg:text-2xl md:text-lg sm:text-sm">
-                      Winners
-                    </TableHead>
-                    <TableHead className="text-slate-200 text-xs lg:text-2xl md:text-lg sm:text-sm">
-                      Win Rate
-                    </TableHead>
-                    <TableHead className="text-slate-200 text-xs lg:text-2xl md:text-lg sm:text-sm">
-                      Best Pair
-                    </TableHead>
-                    <TableHead className="text-slate-200 text-xs lg:text-2xl md:text-lg sm:text-sm">
-                      Worst Pair
-                    </TableHead>
-                  </TableRow>
-                </TableHeader>
-                <TableBody>
-                  {data &&
-                    data.map((data) => (
-                      <TableRow>
-                        <TableCell className="text-xs lg:text-2xl md:text-lg sm:text-sm">
-                          {data.accountid !== undefined ? data.accountid : "0"}
-                        </TableCell>
-                        <TableCell className="text-xs lg:text-2xl md:text-lg sm:text-sm">
-                          {data.totalTrades !== undefined
-                            ? data.totalTrades
-                            : "0"}
-                        </TableCell>
-                        <TableCell className="text-xs lg:text-2xl md:text-lg sm:text-sm">
-                          {data.totalWins !== undefined ? data.totalWins : "0"}
-                        </TableCell>
-                        <TableCell className="text-xs lg:text-2xl md:text-lg sm:text-sm">
-                          <TableCell className="text-xs lg:text-2xl md:text-lg sm:text-sm">
-                            {data.winPercentage !== undefined &&
-                            data.winPercentage !== null
-                              ? `${data.winPercentage}%`
-                              : "0%"}
-                          </TableCell>
-                        </TableCell>
-                        <TableCell className="text-xs lg:text-2xl md:text-lg sm:text-sm">
-                          {data.bestPair !== undefined ? data.bestPair : "N/A"}
-                        </TableCell>
-                        <TableCell className="text-xs lg:text-2xl md:text-lg sm:text-sm">
-                          {data.worstPair !== undefined
-                            ? data.worstPair
-                            : "N/A"}
-                        </TableCell>
-                      </TableRow>
-                    ))}
-                </TableBody>
-              </Table>
-            ) : (
-              "loading"
-            )}
+          <div className="mx-auto">
+            {user ? <HomeTable data={data} /> : "loading"}
           </div>
         </div>
       </div>
